Variable selection.
Here we have checked all the NA values, also we have used cor() function to get the top 10 corelated predictors, but we found that this code only returns predictors that are positively related.

Also we have plotted two pair plots one with predictors from 1 to 10 and later from 11 to 21 and with the help of the co relation plot further in the code we have decided the following predictors.