KNN method & SVM
Here we have added all the models that we have tried, and based on the best RMSE that we have obtained from training the set. Random forest is the model that we have chosen, further under the random forest model code we have predicted the outcome from the test set provided.